Move SplitConcurrencyController to timesharing executor package
