+++ core/trino-main/src/main/java/io/trino/execution/executor/timesharing/TimeSharingTaskHandle.java
@@ -17,7 +17,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.ThreadSafe;
 import com.google.errorprone.annotations.concurrent.GuardedBy;
 import io.airlift.units.Duration;
-import io.trino.execution.SplitConcurrencyController;
 import io.trino.execution.TaskId;
 import io.trino.execution.executor.TaskHandle;
