+++ core/trino-main/src/test/java/io/trino/execution/executor/timesharing/TestSplitConcurrencyController.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.execution;
+package io.trino.execution.executor.timesharing;
 
 import io.airlift.units.Duration;
 import org.junit.jupiter.api.Test;
